Implemented Initial View Restrictions

+++ src/views/layout-employees.vue
@@ -18,6 +18,7 @@
 			<v-dialog
 				max-width="600px"
 				v-model="showAddEmployeeDialog"
+				v-if="isAdmin"
 			>
 				<template v-slot:activator="{ on }">
 					<button
@@ -111,6 +112,7 @@
 			sort-by="name"
 			class="elevation-1"
 		>
+
 			<template v-slot:item.name="{ item, value }">
 				<div class="employee">
 					<router-link
@@ -129,7 +131,10 @@
 			</template>
 
 			<template v-slot:item.employmentStatus="{ item, value }">
-				<span  class="employment_status_label" :class="{ 'fulltime-text': !!value }">{{ !!value ? "Regular" : "Part-time" }}</span>
+				<span
+					class="employment_status_label"
+					:class="{ 'fulltime-text': !!value }"
+				>{{ !!value ? "Regular" : "Part-time" }}</span>
 			</template>
 
 			<template v-slot:item.actions="{ item }">
@@ -137,6 +142,7 @@
 				<button
 					class="action-button"
 					@click="editEmployee(item)"
+					:disabled="!isAdmin"
 				>
 					<svg
 						xmlns="http://www.w3.org/2000/svg"
@@ -147,7 +153,7 @@
 						<path
 							d="M11.089,2.389l2.484,2.484a.269.269,0,0,1,0,.38L7.558,11.269,5,11.553a.536.536,0,0,1-.592-.592L4.694,8.4l6.016-6.016a.269.269,0,0,1,.38,0Zm4.462-.631L14.207.414a1.077,1.077,0,0,0-1.52,0l-.975.975a.269.269,0,0,0,0,.38L14.2,4.254a.269.269,0,0,0,.38,0l.975-.975A1.077,1.077,0,0,0,15.552,1.758ZM10.577,9.633v2.8H1.763V3.623h6.33a.339.339,0,0,0,.234-.1l1.1-1.1a.331.331,0,0,0-.234-.565H1.322A1.322,1.322,0,0,0,0,3.182v9.7A1.322,1.322,0,0,0,1.322,14.2h9.7a1.322,1.322,0,0,0,1.322-1.322V8.531a.331.331,0,0,0-.565-.234l-1.1,1.1A.339.339,0,0,0,10.577,9.633Z"
 							transform="translate(0 -0.1)"
-							fill="#c5c5c5"
+							:class="{ 'disabled-control': !isAdmin }"
 						/>
 					</svg>
 				</button>
@@ -156,6 +162,7 @@
 				<button
 					class="action-button action-delete"
 					@click="deleteEmployee(item)"
+					:disabled="!isAdmin"
 				>
 					<svg
 						xmlns="http://www.w3.org/2000/svg"
@@ -166,7 +173,7 @@
 						<path
 							d="M.875,12.687A1.312,1.312,0,0,0,2.187,14h7.875a1.312,1.312,0,0,0,1.312-1.312V3.5H.875Zm7.437-7a.437.437,0,0,1,.875,0v6.125a.437.437,0,0,1-.875,0Zm-2.625,0a.437.437,0,0,1,.875,0v6.125a.437.437,0,0,1-.875,0Zm-2.625,0a.437.437,0,0,1,.875,0v6.125a.437.437,0,0,1-.875,0ZM11.812.875H8.531L8.274.363A.656.656,0,0,0,7.686,0H4.561a.649.649,0,0,0-.585.364L3.719.875H.437A.437.437,0,0,0,0,1.312v.875a.437.437,0,0,0,.437.437H11.812a.437.437,0,0,0,.437-.437V1.312A.437.437,0,0,0,11.812.875Z"
 							transform="translate(0 0)"
-							fill="#c5c5c5"
+							:class="{ 'disabled-control': !isAdmin }"
 						/>
 					</svg>
 				</button>
@@ -203,6 +210,7 @@ export default {
 			loadingEmployeeDataTable: false,
 			showAddEmployeeDialog: false,
 			departmentCategories: this.$store.state.employees.departments,
+			isAdmin: this.$store.state.user.isAdmin,
 			snackbar: false,
 			text: "Employee Removed.",
 			snackBarTimeOut: 2000
@@ -225,17 +233,23 @@ export default {
 	},
 	methods: {
 		editEmployee(employee) {
-			// TODO: Implement Edit Employee functionality
-			employee;
-			alert("TODO: Implement edit employee functionality");
+			if (this.isAdmin) {
+				// TODO: Implement Edit Employee functionality
+				employee;
+				alert("TODO: Implement edit employee functionality");
+			}
 		},
 		deleteEmployee(employee) {
-			// TODO: show a confirmation dialog to the user before commiting to make an employee as "terminated"
-			this.$store
-				.dispatch("employees/DELETE_EMPLOYEE", employee.id)
-				.then(() => {
-					this.snackbar = true;
-				});
+			if (this.isAdmin) {
+				// TODO: show a confirmation dialog to the user before commiting to make an employee as "terminated"
+				this.$store
+					.dispatch("employees/DELETE_EMPLOYEE", employee.id)
+					.then(() => {
+						this.snackbar = true;
+					});
+			} else {
+				alert("Unauthorized.");
+			}
 		},
 		toggleAddEmployeeDialog() {
 			this.showAddEmployeeDialog = false;
@@ -434,17 +448,23 @@ export default {
 
 .action-button {
 	// background-color: red;
+	fill: #bbbbbb;
 	border: none;
 	padding: 0px 5px;
+	transition: all 0.3s;
 
 	&:hover {
 		cursor: pointer;
-		filter: brightness(80%);
+		filter: brightness(0.6);
 	}
 }
 
-.action-delete:hover path {
-	filter: brightness(100%);
-	fill: rgb(255, 140, 140);
+.action-button:disabled {
+	fill: #cecece;
+	cursor: not-allowed;
+
+	&:hover path {
+		fill: #cecece;
+	}
 }
 </style>
